requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Hardware interfaces (optional - install only if using hardware)
# RPi.GPIO>=0.7.1  # Uncomment for Raspberry Pi GPIO
# adafruit-circuitpython-servokit>=1.3.0  # Uncomment for servo control
# adafruit-circuitpython-ssd1306>=2.12.0  # Uncomment for OLED display
# pynmea2>=1.19.0  # Uncomment for GPS parsing

# AI and ML
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Speech / TTS
pyttsx3>=2.90
gTTS>=2.4.0
playsound>=1.3.0

# Note: Python's `ast` is part of the standard library (no pip install needed).
